♻️ refactor : 메타 데이터 설정
