+++ app/(header-back-title)/report/page.tsx
@@ -1,40 +1,18 @@
-'use client';
 import Header from '@/domains/common/components/header';
-import React, { useState } from 'react';
 import RightArrowIcon from '@/public/icons/arrow-right.svg';
-import ReportTodo from './_components/ReportTodo';
-import ReportDone from './_components/ReportDone';
-import { CartItemState } from '@/domains/cart/types/cartItemType';
+import ReportWrap from './_components/ReportWrap';
+import { Metadata } from 'next';
 
-const TEMPDATA = [
-  {
-    id: '1',
-    title: '이어폰 1',
-    option: '검정색',
-    price: '45,000',
-    image: '',
+export const metadata: Metadata = {
+  title: '리포트 작성하기',
+  description:
+    'Cherry 리포트 작성하기 - 체험한 제품들의 리포트를 작성해 보세요.',
+  icons: {
+    icon: '/icons/Logo.svg',
   },
-  {
-    id: '2',
-    title: '이어폰 2',
-    option: '하늘색',
-    price: '45,000',
-    image: '',
-  },
-  {
-    id: '3',
-    title: '이어폰 3',
-    option: '흰색',
-    price: '45,000',
-    image: '',
-  },
-];
+};
 
 export default function Page() {
-  const [selectTab, setSelectTab] = useState('Todo');
-  const [todoData, setTodoData] = useState<CartItemState[]>(TEMPDATA);
-  const [doneData, setDoneData] = useState<CartItemState[]>(TEMPDATA);
-
   return (
     <div className="h-screen overflow-hidden bg-Fill-99">
       <Header
@@ -85,37 +63,7 @@ export default function Page() {
             <RightArrowIcon className="w-4 h-4 fill-Fill-20" />
           </a>
         </div>
-        <div className="self-stretch px-4 bg-Static-White border-b border-Line-Subtler flex gap-2 justify-start items-center">
-          <button
-            type="button"
-            className={`p-3 cursor-pointer ${
-              selectTab === 'Todo'
-                ? 'border-b-2 border-Primary-Normal title2-b text-Label-Normal'
-                : 'body1-m text-Label-Normal'
-            }`}
-            onClick={() => setSelectTab('Todo')}
-          >
-            작성 가능한 리포트
-          </button>
-          <button
-            type="button"
-            className={`p-3 cursor-pointer ${
-              selectTab === 'Done'
-                ? 'border-b-2 border-Primary-Normal title2-b text-Label-Normal'
-                : 'body1-m text-Label-Normal'
-            }`}
-            onClick={() => setSelectTab('Done')}
-          >
-            작성한 리포트
-          </button>
-        </div>
-        <div className="h-[calc(100vh-250px)] overflow-y-auto no-scrollbar">
-          {selectTab === 'Todo' ? (
-            <ReportTodo todoData={todoData} />
-          ) : (
-            <ReportDone doneData={doneData} />
-          )}
-        </div>
+        <ReportWrap />
       </main>
     </div>
   );
