+++ app/(header-back-title-cart)/products/[id]/_components/ProductWrap.tsx
@@ -1,0 +1,62 @@
+'use client';
+import React, { useEffect, useState } from 'react';
+import ProductDetail from './ProductDetail';
+import ProductDetailInfo from './ProductDetailInfo';
+import ProductDetailGuide from './ProductDetailGuide';
+import FooterBtn from './FooterBtn';
+import ToastComponent from '@/domains/common/components/ToastComponent';
+import ModalComponent from '@/domains/common/components/ModalComponent';
+import { useToastStore } from '@/domains/common/store/toastStore';
+import { getAPI } from '@/domains/common/api';
+
+interface ProductDetailType {
+  name: string;
+  price: number;
+  description: string;
+}
+
+export default function ProductWrap({ id }: { id: string }) {
+  const [detailData, setDetailData] = useState<ProductDetailType | null>(null);
+  const [showOptions, setShowOptions] = useState(false);
+  const [showModal, setShowModal] = useState(false);
+  const { show } = useToastStore.getState();
+
+  useEffect(() => {
+    const getData = async () => {
+      const res = await getAPI<ProductDetailType>(`/productDetail?${id}`);
+      setDetailData(res);
+    };
+
+    // TODO : api 연결 후 주석 해제
+    // getData();
+  }, [id]);
+
+  return (
+    <>
+      <div className="pb-3 flex flex-col gap-3 overflow-y-scroll h-[calc(100vh-135px)] no-scrollbar">
+        <ProductDetail />
+        <ProductDetailInfo />
+        <ProductDetailGuide />
+      </div>
+      <FooterBtn
+        id={id as string}
+        showOptions={showOptions}
+        setShowOptions={setShowOptions}
+        showModal={showModal}
+        setShowModal={setShowModal}
+      />
+      {showOptions && (
+        <div className="absolute inset-0 bg-black opacity-30 z-40" />
+      )}
+
+      {show && <ToastComponent />}
+      {showModal && (
+        <ModalComponent
+          title="옵션을 선택해 주세요!"
+          content="1개 이상의 옵션을 선택해야 합니다."
+          onConfirm={() => setShowModal(false)}
+        />
+      )}
+    </>
+  );
+}
